requirements.txt
requests 
beautifulsoup4 
pdf2image 
pytesseract 
openai 
streamlit 
PyMuPDF 
numpy
fastapi 
uvicorn 
openpyxl
python-dotenv
together
xlsxwriter
